XSLT-process result XML manually, don't just reference stylesheet.
This has two advantages.  On the one hand we write fully valid HTML
protocols if our user hits the store button.  On the other hand we
have no problem with the XSS protection if run in a browser.
Browsers deny access to XSL files loaded via http if the document
to process is provided via data protocol (at least Firefox does).